Adding Source to Commit
- moved commit type models into single folder
- update commit call
- moved enums to own file
- state fix
