+++ FitpaySDK/Rest/Models/Commits/CommitMetrics.swift
@@ -1,11 +1,4 @@
-
-public enum SyncInitiator: String, Serializable {
-    case platform = "PLATFORM"
-    case notification = "NOTIFICATION"
-    case webHook = "WEB_HOOK"
-    case eventStream = "EVENT_STREAM"
-    case notDefined = "NOT DEFINED"
-}
+import Foundation
 
 open class CommitMetrics: Serializable {
     public var syncId: String?
