+++ FitpaySDK/Rest/Models/Commits/Enums/APDUPackageResponseState.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public enum APDUPackageResponseState: String, Serializable {
+    case processed    = "PROCESSED"
+    case failed       = "FAILED"
+    case error        = "ERROR"
+    case expired      = "EXPIRED"
+    case notProcessed = "NOT_PROCESSED"
+}
