+++ FitpaySDK/Rest/Models/Commits/Enums/SourceType.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public enum SourceType: String, Serializable {
+    case deviceApi = "device"
+    case tsmGatewayApi = "tsm"
+    case gAndDIntegrationApi = "gi_de"
+    case mdesGatewayApi = "mdes"
+    case discoverGatewayApi = "discover"
+    case vdepGatewayApi = "vdep"
+}
